pyproject.toml
[project]
name = "agentic-phidata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckduckgo-search>=7.2.1",
    "groq>=0.15.0",
    "phidata>=2.7.8",
    "yfinance>=0.2.52",
]
